implemented components and react hook form

+++ src/App.jsx
@@ -1,27 +1,31 @@
 import { useState, useEffect } from "react";
 import "./App.css";
-import Box from "@mui/material/Box";
-import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
-import Button from "@mui/material/Button";
 import { v4 as uuidv4 } from "uuid";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import Checkbox from "@mui/material/Checkbox";
-import IconButton from "@mui/material/IconButton";
-import DeleteIcon from "@mui/icons-material/Delete";
 import { ToastContainer, toast } from "react-toastify";
-import EditIcon from "@mui/icons-material/Edit";
-import CircularProgress from "@mui/material/CircularProgress";
 import { ShowToast } from "./utl/Showtoast";
+import Todoform from "./components/Todoform";
+import Todolist from "./components/Todolist";
+import { Form, useForm } from "react-hook-form";
+
 function App() {
-  const [form, setForm] = useState({
-    task: "",
-    isCompleted: false,
-  });
+  const [form, setForm] = useState();
   const [todos, settodos] = useState([]);
   const [loading, setloading] = useState(false);
+
+  // react hook form
+
+  const {
+    register,
+    handleSubmit,
+    watch,
+    setValue,
+    formState: { errors },
+  } = useForm();
+
+  function handleDataFromTodoForm(setValue) {
+    setForm(() => setValue);
+  }
 
   useEffect(() => {
     async function fetchtTodos() {
@@ -38,71 +42,14 @@ function App() {
     fetchtTodos();
   }, []);
 
-  function handleForm(event) {
-    setForm({
-      ...form,
-      [event.target.name]: event.target.value,
-    });
-  }
-
-  async function handleKeyDown(event) {
-    if (event.key === "Enter") {
-      const editedTodo = todos.filter((todo) => {
-        return todo.id !== form.id;
-      });
-      const newTodo = { ...form, id: uuidv4() };
-      if (newTodo.task) {
-        const updatedTodos = [...editedTodo, newTodo];
-        settodos(updatedTodos);
-        setForm({
-          ...form,
-          task: "",
-        });
-
-        try {
-          await fetch(import.meta.env.VITE_SERVER_URI, {
-            method: "DELETE",
-            body: JSON.stringify({
-              id: form.id,
-            }),
-            headers: {
-              "Content-type": "application/json; charset=UTF-8",
-            },
-          });
-        } catch (error) {
-          console.log("An error occured,", error.message);
-        }
-
-        try {
-          await fetch(import.meta.env.VITE_SERVER_URI, {
-            method: "POST",
-            body: JSON.stringify({
-              task: form.task,
-              isCompleted: form.isCompleted,
-              id: updatedTodos[updatedTodos.length - 1].id,
-            }),
-            headers: {
-              "Content-type": "application/json; charset=UTF-8",
-            },
-          });
-        } catch (error) {
-          console.log("An error occured,", error.message);
-        }
-        ShowToast("info", "Your todo is saved");
-      } else {
-        ShowToast("warn", "Please add some value");
-      }
-    }
-  }
-
-  async function saveTodo() {
-    const editedArray = todos.filter((todo) => {
-      return todo.id !== form.id;
-    });
-    try {
+  async function saveTodo(task) {
+    if (form) {
+     const editedTodo = todos.filter((todo)=>{ return todo.id !== form.id})
+     settodos(editedTodo);     
+     try {
       await fetch(import.meta.env.VITE_SERVER_URI, {
         method: "DELETE",
-        body: JSON.stringify({ id: form.id }),
+        body: form,
         headers: {
           "Content-type": "application/json; charset=UTF-8",
         },
@@ -110,26 +57,14 @@ function App() {
     } catch (error) {
       console.log("An error occured,", error.message);
     }
-    const newTodo = { ...form, isCompleted: false, id: uuidv4() };
-    if (!newTodo.task) {
-      ShowToast("warn", "Please add some value");
-      return;
     }
-    const updatedTodos = [...editedArray, newTodo];
-    settodos(updatedTodos);
-    setForm({
-      ...form,
-      task: "",
-    });
-
+    setForm(task);
+    const dataWithid = { ...task, isCompleted: false, id: uuidv4() };
+    settodos((prevTodos) => [...prevTodos, dataWithid]);
     try {
       await fetch(import.meta.env.VITE_SERVER_URI, {
         method: "POST",
-        body: JSON.stringify({
-          task: form.task,
-          isCompleted: form.isCompleted,
-          id: updatedTodos[updatedTodos.length - 1].id,
-        }),
+        body: JSON.stringify(dataWithid),
         headers: {
           "Content-type": "application/json; charset=UTF-8",
         },
@@ -194,9 +129,9 @@ function App() {
 
   async function handleEdit(myid) {
     let editArray = todos.filter((todo) => todo.id === myid)[0];
-    setForm({ ...form, task: editArray.task, id: editArray.id });
+    setValue("task", editArray.task);
+    setForm(editArray)
   }
-
   return (
     <>
       <ToastContainer
@@ -214,118 +149,25 @@ function App() {
       <Typography variant="h2" gutterBottom>
         Neeraj's Todos
       </Typography>
-      <Box
-        sx={{
-          flexDirection: "row",
-          display: "flex",
-          justifyContent: "center",
-          gap: "10px",
-        }}
-      >
-        <TextField
-          onChange={handleForm}
-          onKeyDown={handleKeyDown}
-          name="task"
-          value={form.task}
-          id="outlined-basic"
-          variant="outlined"
-          size="small"
-          label="My Task"
-        />
-        <Button onClick={saveTodo} variant="outlined" size="medium">
-          Add
-        </Button>
-      </Box>
-      <Box
-        sx={{
-          flexDirection: "column",
-          display: "flex",
-          justifyContent: "center",
-          gap: "0",
-          alignItems: "center",
-        }}
-      >
-        {loading && (
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "center",
-              marginTop: "40px",
-            }}
-          >
-            <CircularProgress />
-          </Box>
-        )}
-        {todos.length == 0 && loading === false ? (
-          <Box
-            sx={(theme) => ({
-              p: 1,
-              color: "grey.800",
-              fontSize: "0.575rem",
-            })}
-          >
-            <Typography variant="h4" component="h2">
-              Please Add new tasks
-            </Typography>
-          </Box>
-        ) : null}
-        {todos.length !== 0 && loading === false
-          ? todos.map((todo) => {
-              return (
-                <List
-                  key={todo.id}
-                  dense
-                  sx={{
-                    width: "100%",
-                    maxWidth: "768px",
-                    bgcolor: "background.paper",
-                  }}
-                >
-                  <ListItem
-                    secondaryAction={
-                      <>
-                        <IconButton aria-label="delete" size="large">
-                          <EditIcon
-                            fontSize="inherit"
-                            onClick={() => handleEdit(todo.id)}
-                          />
-                        </IconButton>
-
-                        <Checkbox
-                          edge="end"
-                          onChange={() => markAsDone(todo.id, todo.isCompleted)}
-                          checked={todo.isCompleted ? "checked" : ""}
-                          inputProps=""
-                        />
-                        <IconButton aria-label="delete" size="large">
-                          <DeleteIcon
-                            fontSize="inherit"
-                            onClick={() => deleteTodo(todo.id)}
-                          />
-                        </IconButton>
-                      </>
-                    }
-                    disablePadding
-                  >
-                    <ListItemButton>
-                      <Typography
-                        variant="h5"
-                        component="h2"
-                        style={{
-                          textDecoration: todo.isCompleted
-                            ? "line-through"
-                            : "none",
-                        }}
-                      >
-                        {todo.task}
-                      </Typography>
-                    </ListItemButton>
-                  </ListItem>
-                </List>
-              );
-            })
-          : null}
-      </Box>
+      <Todoform
+        sendDataToParent={handleDataFromTodoForm}
+        saveTodo={saveTodo}
+        settodos={settodos}
+        form={form}
+        register={register}
+        handleSubmit={handleSubmit}
+        watch={watch}
+        setValue={setValue}
+        errors={errors}
+        setForm={setForm}
+      />
+      <Todolist
+        loading={loading}
+        todos={todos}
+        handleEdit={handleEdit}
+        markAsDone={markAsDone}
+        deleteTodo={deleteTodo}
+      />
     </>
   );
 }
